+++ src/test/java/io/github/mat3e/HelloServiceTest.java
@@ -31,7 +31,7 @@ public class HelloServiceTest {
     public void test_prepareGreeting_nullLang_returnsGreetingWithFallbackIdLang() {
         var mockRepository = new LangRepository() {
             @Override
-            Optional<Lang> findByID(Long id) {
+            Optional<Lang> findByID(Integer id) {
                 if (id.equals(HelloService.FALLBACK_LANG.getId())) {
                     return Optional.of(new Lang(null, FALLBACK_ID_WELCOME, null));
                 }
@@ -47,7 +47,7 @@ public class HelloServiceTest {
     public void test_prepareGreeting_textLang_returnsGreetingWithFallbackIdLang() {
         var mockRepository = new LangRepository() {
             @Override
-            Optional<Lang> findByID(Long id) {
+            Optional<Lang> findByID(Integer id) {
                 if (id.equals(HelloService.FALLBACK_LANG.getId())) {
                     return Optional.of(new Lang(null, FALLBACK_ID_WELCOME, null));
                 }
@@ -63,7 +63,7 @@ public class HelloServiceTest {
     public void test_prepareGreeting_nonExistingLang_returnsGreetingWithFallbackLang() {
         var mockRepository = new LangRepository() {
             @Override
-            Optional<Lang> findByID(Long id) {
+            Optional<Lang> findByID(Integer id) {
                 return Optional.empty();
             }
         };
@@ -75,7 +75,7 @@ public class HelloServiceTest {
     private LangRepository alwaysReturningHelloRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findByID(Long id) {
+            Optional<Lang> findByID(Integer id) {
                 return Optional.of(new Lang(null, "Hello", null));
             }
         };
